Split target analysis into separate classes.
