+++ nonbonded/library/factories/analysis/optimization.py
@@ -1,56 +1,31 @@
 import errno
-import json
 import logging
 import os
 from collections import defaultdict
 from glob import glob
-from typing import Optional
 
 from nonbonded.library.factories.analysis import AnalysisFactory
-from nonbonded.library.models.authors import Author
-from nonbonded.library.models.datasets import Component, DataSet
+from nonbonded.library.factories.analysis.targets.evaluator import (
+    EvaluatorAnalysisFactory,
+)
+from nonbonded.library.factories.analysis.targets.recharge import (
+    RechargeAnalysisFactory,
+)
 from nonbonded.library.models.forcefield import ForceField
 from nonbonded.library.models.projects import Optimization
-from nonbonded.library.models.results import (
-    DataSetResult,
-    EvaluatorTargetResult,
-    OptimizationResult,
-    RechargeTargetResult,
-    Statistic,
-)
+from nonbonded.library.models.results import OptimizationResult
 from nonbonded.library.models.targets import EvaluatorTarget, RechargeTarget
-from nonbonded.library.statistics.statistics import StatisticType, bootstrap_residuals
-from nonbonded.library.utilities.checkmol import components_to_categories
-from nonbonded.library.utilities.migration import reindex_results
 from nonbonded.library.utilities.provenance import summarise_current_versions
 
 logger = logging.getLogger(__name__)
 
+_TARGET_FACTORIES = {
+    EvaluatorTarget: EvaluatorAnalysisFactory,
+    RechargeTarget: RechargeAnalysisFactory,
+}
+
 
 class OptimizationAnalysisFactory(AnalysisFactory):
-    @classmethod
-    def _read_objective_function(cls, target_directory) -> float:
-        """Reads the value of the objective function from a ForceBalance
-        nifty file stored in an iteration output directory
-
-        Parameters
-        ----------
-        target_directory
-            The directory which contains the nifty file.
-
-        Returns
-        -------
-            The value of the objective function.
-        """
-
-        from forcebalance.nifty import lp_load
-
-        # Extract the value of this iterations objective function
-        objective_file_path = os.path.join(target_directory, "objective.p")
-        objective_statistics = lp_load(objective_file_path)
-
-        return objective_statistics["X"]
-
     @classmethod
     def _load_refit_force_field(cls) -> ForceField:
         """Load in the refit force field."""
@@ -74,114 +49,6 @@ class OptimizationAnalysisFactory(AnalysisFactory):
         refit_force_field = ForceField.from_openff(refit_force_field_off)
 
         return refit_force_field
-
-    @classmethod
-    def _analyze_evaluator_target(
-        cls,
-        optimization: Optimization,
-        target: EvaluatorTarget,
-        target_directory: str,
-        reindex: bool,
-    ) -> Optional[EvaluatorTargetResult]:
-
-        from openff.evaluator.client import RequestResult
-        from openff.evaluator.datasets import PhysicalPropertyDataSet
-
-        results_path = os.path.join(target_directory, "results.json")
-
-        if not os.path.isfile(results_path):
-            return None
-
-        # Load the reference data set
-        reference_data_set = DataSet.from_pandas(
-            PhysicalPropertyDataSet.from_json(
-                os.path.join("targets", target.id, "training-set.json")
-            ).to_pandas(),
-            identifier="empty",
-            description="empty",
-            authors=[Author(name="empty", email="email@email.com", institute="empty")],
-        )
-
-        results = RequestResult.from_json(results_path)
-
-        if reindex:
-            results = reindex_results(results, reference_data_set)
-
-        estimated_data_set = results.estimated_properties
-
-        # Generate statistics about each iteration.
-        data_set_result = DataSetResult.from_evaluator(
-            reference_data_set=reference_data_set,
-            estimated_data_set=estimated_data_set,
-            analysis_environments=optimization.analysis_environments,
-            statistic_types=[StatisticType.RMSE],
-        )
-
-        objective_function = cls._read_objective_function(target_directory)
-
-        return EvaluatorTargetResult(
-            objective_function=target.weight * objective_function,
-            statistic_entries=data_set_result.statistic_entries,
-        )
-
-    @classmethod
-    def _analyze_recharge_target(
-        cls,
-        optimization: Optimization,
-        target: RechargeTarget,
-        target_directory: str,
-    ) -> Optional[RechargeTargetResult]:
-
-        residuals_path = os.path.join(target_directory, "residuals.json")
-
-        if not os.path.isfile(residuals_path):
-            return None
-
-        # Load in the residuals
-        with open(residuals_path) as file:
-            squared_residuals = json.load(file)
-
-        # Categorize the smiles
-        smiles_per_category = defaultdict(list)
-
-        smiles_per_category[None] = [*squared_residuals]
-
-        for smiles in squared_residuals:
-
-            categories = components_to_categories(
-                [Component(smiles=smiles, mole_fraction=0.0, exact_amount=1)],
-                optimization.analysis_environments,
-            )
-
-            for category in categories:
-                smiles_per_category[category].append(smiles)
-
-        # Compute RMSE statistics for this target.
-        statistic_entries = []
-
-        for category in smiles_per_category:
-
-            category_residuals = [
-                squared_residuals[smiles] for smiles in smiles_per_category[category]
-            ]
-
-            rmse, rmse_std, rmse_ci = bootstrap_residuals(category_residuals)
-
-            statistic_entry = Statistic(
-                statistic_type=StatisticType.RMSE,
-                category=category,
-                value=rmse,
-                lower_95_ci=rmse_ci[0],
-                upper_95_ci=rmse_ci[1],
-            )
-            statistic_entries.append(statistic_entry)
-
-        objective_function = cls._read_objective_function(target_directory)
-
-        return RechargeTargetResult(
-            objective_function=target.weight * objective_function,
-            statistic_entries=statistic_entries,
-        )
 
     @classmethod
     def analyze(cls, reindex):
@@ -230,16 +97,23 @@ class OptimizationAnalysisFactory(AnalysisFactory):
                 )
 
                 # Analyse the target
-                if isinstance(target, EvaluatorTarget):
-                    target_result = cls._analyze_evaluator_target(
-                        optimization, target, iteration_directory, reindex
-                    )
-                elif isinstance(target, RechargeTarget):
-                    target_result = cls._analyze_recharge_target(
-                        optimization, target, iteration_directory
-                    )
-                else:
-                    raise NotImplementedError()
+                target_analyzer = _TARGET_FACTORIES.get(target.__class__, None)
+
+                if target_analyzer is None:
+                    raise NotImplementedError
+
+                target_analyzer_kwargs = {}
+
+                if isinstance(target_analyzer, EvaluatorAnalysisFactory):
+                    target_analyzer_kwargs["reindex"] = reindex
+
+                target_result = target_analyzer.analyze(
+                    optimization,
+                    target,
+                    os.path.join("targets", target.id),
+                    iteration_directory,
+                    **target_analyzer_kwargs,
+                )
 
                 if target_result is None:
