+++ nonbonded/library/factories/analysis/targets/__init__.py
@@ -1,0 +1,3 @@
+from nonbonded.library.factories.analysis.targets.targets import TargetAnalysisFactory
+
+__all__ = [TargetAnalysisFactory]
